Share probabilityChart component between scalar and categorical market
